refac: widgets - post-manager-hearder 분리
