+++ src/widgets/index.ts
@@ -3,6 +3,7 @@ export { Footer } from "@widgets/footer"
 export { PostTable } from "@widgets/post-table"
 export { PostFilters } from "@widgets/post-filters"
 export { PostDetailDialog } from "@widgets/post-detail-dialog"
+export { PostManagerHeader } from "@widgets/post-manager-header"
 export { Pagination } from "@widgets/pagination"
 export { UserDialog } from "@widgets/user-dialog"
 export { CommentFormDialog } from "@widgets/comment-form-dialog"
