+++ src/widgets/post-manager-header/index.ts
@@ -1,0 +1,1 @@
+export { PostManagerHeader } from "./ui"
