+++ src/widgets/post-manager-header/ui/index.ts
@@ -1,0 +1,1 @@
+export { PostManagerHeader } from "./post-manager-header"
